SF Bug #1023229
Relatively simple fix to a complex, that changes the weekly code to render to the window of $next_date_range -> $next_date_range + $interval weeks. Took me a while to wrap my head around this one - I knew what the solution was, but wasn't sure if it handles everything. Still not 100% sure about the edge cases, though.